adding authorization checks

+++ rest-api/controllers/feed.js
@@ -128,6 +128,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if(post.creator.toString() !== req.userId){
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       if(imageUrl !== post.imageUrl){
         clearImage(post.imageUrl);  
       }
@@ -156,7 +161,11 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      //check log in user
+      if(post.creator.toString() !== req.userId){
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
